Fixing project importing
