+++ .gitignore
@@ -7,6 +7,11 @@ __pycache__/
 # C extensions
 *.so
 
+# Please dont remove this again
+/.DS_Store
+*/.DS_Store
+.DS_Store
+
 # Distribution / packaging
 .Python
 build/
