+++ src/email_analysis/pyproject.toml
@@ -1,14 +1,10 @@
 [project]
-name = "Email Analysis"
+name = "email_analysis"
 version = "0.1.0"
 description = "Email Analyzer Protocol"
 readme = "README.md"
 requires-python = ">=3.11"
 
-  
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
-
-[tool.uv.sources]
-message = { workspace = true }
