+++ src/email_analysis/src/email_analysis/__init__.py
