+++ src/email_analysis_impl/README.md
@@ -1,0 +1,18 @@
+# Inbox Client Protocol
+
+## Overview
+
+
+## Scope
+
+
+### Out of Scope
+
+
+## Protocol Definition
+
+
+### `analyze() -> dict`
+
+
+## Usage
