+++ src/email_analysis_impl/src/email_analysis_impl/_impl.py
