+++ src/inbox_client/pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "inbox-client-api"
+name = "inbox_client"
 version = "0.1.0"
 description = "Inbox Client Protocol"
 readme = "README.md"
