+++ src/inbox_client_impl/src/inbox_client_impl/__init__.py
@@ -1,7 +1,46 @@
+from typing import ClassVar, Iterator
+from datetime import datetime
+from googleapiclient.discovery import build
+from google.auth.transport.requests import Request
+from google_auth_oauthlib.flow import InstalledAppFlow
+from google.oauth2.credentials import Credentials
+import base64
+import os.path
 import inbox_client_api
+import message
 
-from . import _impl
+'''
+you can put all this links in constants folder and fetch from it to make it cleaner
 
-# Dependency Injection of this implementation into the API
-#
-inbox_client_api.get_client = lambda: _impl.Client()
+i noticed the implementation of the other methods such as send delete and mark as read in the test and readme files but the impl is missing in the inbox_client.
+
+
+'''
+
+class GmailClient(inbox_client_api.Client):
+    SCOPES: ClassVar[list[str]] = ["https://www.googleapis.com/auth/gmail.readonly"]
+
+    def __init__(self):
+        creds = None
+        if os.path.exists('token.json'):
+            creds = Credentials.from_authorized_user_file('token.json', self.SCOPES)
+        if not creds or not creds.valid:
+            if creds and creds.expired and creds.refresh_token:
+                creds.refresh(Request())
+            else:
+                flow = InstalledAppFlow.from_client_secrets_file('credentials.json', self.SCOPES)
+                creds = flow.run_local_server(port=0)
+            with open('token.json', 'w') as token:
+                token.write(creds.to_json())
+        self.service = build('gmail', 'v1', credentials=creds)
+    
+    def get_messages(self) -> Iterator[message.Message]:
+        results = self.service.users().messages().list(userId='me', maxResults=10).execute()
+        messages = results.get('messages', [])
+
+        for msg in messages:
+            msg_data = self.service.users().messages().get(userId='me', id=msg['id'], format='raw').execute()
+            yield message.Message(msg_data)
+
+def get_client() -> GmailClient:
+    return GmailClient()
