+++ tests/email_analysis/test_llm_analyzer.py
@@ -1,5 +1,5 @@
 import pytest
-from src.email_analysis_impl._impl import LLMAnalyzer
+from email_analysis_impl.src.email_analysis_impl._impl import LLMAnalyzer
 
 
 def test_analyze_spam_and_importance():
